add e2e squeezenet
