+++ .travis.yml
@@ -5,10 +5,18 @@ dist: trusty
 node_js:
   - '8.9.1'
 
+addons:
+  chrome: stable
+  firefox: stable
+
 services:
   - docker
 
 before_install:
+  # We need a display for the browsers
+  - export DISPLAY=:99.0
+  - sh -e /etc/init.d/xvfb start
+  - sleep 3
 
   # Pull the official caffe docker image
   - docker pull bvlc/caffe:cpu
@@ -25,3 +33,4 @@ install:
 
 script:
   - npm run build
+  - npm run test-e2e-travis
