+++ karma.e2e.conf.js
@@ -4,6 +4,7 @@ module.exports = function(config) {
     files: [
       { pattern: "src/**/*.ts" },
       { pattern: "test/**/*.ts" },
+      { pattern: "test/**/*.jpg", watched: false, included: false },
       { pattern: "test/**/model/**/*", watched: false, included: false },
       { pattern: "test/**/activations/**/*", watched: false, included: false }
     ],
@@ -12,6 +13,10 @@ module.exports = function(config) {
       tsconfig: 'tsconfig.json'
     },
     reporters: ["progress", "karma-typescript"],
-    browsers: ["Chrome"]
+    browsers: ["Chrome"],
+
+    // Wait maximal 5 min for a single test.
+    // this is needed for CPU pass through a model
+    browserNoActivityTimeout: 1000*60*5
   });
 };
