+++ test/e2e/run.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Run caffe inference
+sh "$(dirname $0)/squeezenet/run.sh"
+
+# Run Karma test runner
+node_modules/.bin/karma start $@ karma.e2e.conf.js
