+++ test/e2e/squeezenet/cat.jpg
